arcreco/requirements.txt
asgiref==3.3.1
boto3==1.16.56
botocore==1.19.56
certifi==2020.12.5
chardet==4.0.0
Django==2.2
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.0
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
mysqlclient==2.0.3
numpy==1.19.5
openpyxl==3.0.6
pandas==0.25.0
Pillow==8.1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-environ==0.4.54
pytz==2020.5
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
xlrd==2.0.1
